refactor: Move IP matching logic to dedicated package
Moves the IP matching logic from the `modules` package to a new
`matchers/ip` package. This change improves code organization and
clarifies the purpose of the different components.  It also adds the
`maxminddb-golang` dependency to support IP matching against MaxMind
databases.

The `whitelist` module is also moved to the `matchers` package.  Tests
are updated to reflect the new package structure.